Implement rudimentary command forwarding
<prefix> commands in Discord will now send to Minecraft and
execute as expected. Works with any command added by any plugin
that can be executed as server.

Command forwarding is not very robust, does not give feedback,
and will attempt to send a command before checking if it exists or not.
Will work on fixing those in future commits.

Also removed some console spam.