Add custom Markdown parser
A custom InlineParser replacement has been added, based on Flexmark's InlineParserImpl.
This new parser only concerns itself with the type of Markdown constructs that Discord
is capable of creating, ignoring everything else in the CommonMark spec.

A rudimentary NodeVisitor has also been implemented to render the AST in Minecraft, but
it does not work yet...